A first pass at vim documentation
This is mostly the README shifted over to the `help.vim` format. I fixed
some typos, and I made some minor changes. In addition, I added one
large example (on using the toggle commands in autocommands) and one
whole section (on how to help vim know that both “isn't” and “isn’t” are
spelled correctly). I also added the license directly into the help
(which I think is common), and I updated the copyright date.

I will continue to edit and push further changes as additional commits
to this branch.